requirements.txt
fastapi
uvicorn
pydantic
motor
python-dotenv
pycryptodome
pyjwt
pytest
httpx